Check Window termination to fix crash on ProgressBar property access after window closed
